Set the namespace 'api' for the prepare of the webpacker
