+++ app/controllers/api/base_controller.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Api
+  class BaseController < ApplicationController
+    before_action :authenticate_user!
+    before_action :set_locale
+
+    private
+
+    def set_locale
+      profile = current_user.profile || current_user.build_profile
+      I18n.locale = profile.locale
+    end
+  end
+end
